[lexer] New handling of compiler directives
While this is not a preprocessor and we don't expect to encounter
directives, if we do we need to grab all the characters that 'belongs' to
the directive and remove them from the character stream. We do this to
avoid interpreting the directive text as regular Verilog and to let the
parser handle these tokens as an atomic block.

There are a number of compiler directives defined by the standard. For
these, we eat all characters until the next newline. We continue with the
next line if the newline character is escaped (as is allowed for `define).
However, if we don't recognize the directive, we just grab the first word
and optionally a comma-separated list of arguments enclosed in parenthesis.
The assumption is that this is a text macro usage.